# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"JsonValue": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					},
					{
						"$ref": "#/components/schemas/JsonObject"
					},
					{
						"$ref": "#/components/schemas/JsonArray"
					}
				],
				"nullable": true,
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches any valid JSON value."
			},
			"JsonObject": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/JsonValue"
				},
				"type": "object",
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches a JSON object.\r\nThis type can be useful to enforce some input to be JSON-compatible or as a super-type to be extended from."
			},
			"JsonArray": {
				"description": "From https://github.com/sindresorhus/type-fest/\r\nMatches a JSON array.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.TipoUsuario": {
				"type": "string",
				"enum": [
					"INDIVIDUAL",
					"EMPRESARIAL"
				]
			},
			"LoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"requiereCompletarPerfil": {
						"type": "boolean"
					},
					"user": {
						"properties": {
							"empresaId": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"perfilCompleto": {
								"type": "boolean"
							},
							"rolEquipo": {
								"type": "string"
							},
							"tipoUsuario": {
								"type": "string",
								"nullable": true
							},
							"rol": {
								"type": "string"
							},
							"correo": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"idUsuario": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"empresaId",
							"perfilCompleto",
							"rolEquipo",
							"tipoUsuario",
							"rol",
							"correo",
							"username",
							"idUsuario"
						],
						"type": "object"
					}
				},
				"required": [
					"token",
					"requiereCompletarPerfil",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"correo",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SolicitudResetDTO": {
				"properties": {
					"correo": {
						"type": "string"
					}
				},
				"required": [
					"correo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfirmacionResetDTO": {
				"properties": {
					"token": {
						"type": "string"
					},
					"nuevaContrasena": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"nuevaContrasena"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/nutriscan": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Registro creado",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"isTest": {
													"type": "boolean"
												},
												"actualizadoEn": {
													"type": "string",
													"format": "date-time"
												},
												"fechaAnalisis": {
													"type": "string",
													"format": "date-time"
												},
												"tipoAnalisis": {
													"type": "string"
												},
												"respuesta": {
													"$ref": "#/components/schemas/JsonValue"
												},
												"consulta": {
													"type": "string"
												},
												"esAlimento": {
													"type": "boolean"
												},
												"usuarioId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"isTest",
												"actualizadoEn",
												"fechaAnalisis",
												"tipoAnalisis",
												"respuesta",
												"consulta",
												"esAlimento",
												"usuarioId",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					},
					"403": {
						"description": "Acceso denegado"
					}
				},
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"isTest": {
														"type": "boolean"
													},
													"actualizadoEn": {
														"type": "string",
														"format": "date-time"
													},
													"fechaAnalisis": {
														"type": "string",
														"format": "date-time"
													},
													"tipoAnalisis": {
														"type": "string"
													},
													"respuesta": {
														"$ref": "#/components/schemas/JsonValue"
													},
													"consulta": {
														"type": "string"
													},
													"esAlimento": {
														"type": "boolean"
													},
													"usuarioId": {
														"type": "number",
														"format": "double"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"isTest",
													"actualizadoEn",
													"fechaAnalisis",
													"tipoAnalisis",
													"respuesta",
													"consulta",
													"esAlimento",
													"usuarioId",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/nutriscan/usuario/{usuarioId}": {
			"get": {
				"operationId": "FindByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"allOf": [
													{
														"properties": {
															"usuario": {
																"properties": {
																	"tipoUsuario": {
																		"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
																	},
																	"nombreCompleto": {
																		"type": "string"
																	}
																},
																"required": [
																	"tipoUsuario",
																	"nombreCompleto"
																],
																"type": "object"
															}
														},
														"required": [
															"usuario"
														],
														"type": "object"
													},
													{
														"properties": {
															"isTest": {
																"type": "boolean"
															},
															"actualizadoEn": {
																"type": "string",
																"format": "date-time"
															},
															"fechaAnalisis": {
																"type": "string",
																"format": "date-time"
															},
															"tipoAnalisis": {
																"type": "string"
															},
															"respuesta": {
																"$ref": "#/components/schemas/JsonValue"
															},
															"consulta": {
																"type": "string"
															},
															"esAlimento": {
																"type": "boolean"
															},
															"usuarioId": {
																"type": "number",
																"format": "double"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"isTest",
															"actualizadoEn",
															"fechaAnalisis",
															"tipoAnalisis",
															"respuesta",
															"consulta",
															"esAlimento",
															"usuarioId",
															"id"
														],
														"type": "object"
													}
												]
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "usuarioId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/nutriscan/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"isTest": {
													"type": "boolean"
												},
												"actualizadoEn": {
													"type": "string",
													"format": "date-time"
												},
												"fechaAnalisis": {
													"type": "string",
													"format": "date-time"
												},
												"tipoAnalisis": {
													"type": "string"
												},
												"respuesta": {
													"$ref": "#/components/schemas/JsonValue"
												},
												"consulta": {
													"type": "string"
												},
												"esAlimento": {
													"type": "boolean"
												},
												"usuarioId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"isTest",
												"actualizadoEn",
												"fechaAnalisis",
												"tipoAnalisis",
												"respuesta",
												"consulta",
												"esAlimento",
												"usuarioId",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					}
				},
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"NutriScan"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/solicitar-reset": {
			"post": {
				"operationId": "SolicitarReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SolicitudResetDTO"
							}
						}
					}
				}
			}
		},
		"/auth/confirmar-reset": {
			"post": {
				"operationId": "ConfirmarReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfirmacionResetDTO"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}